Fix update of boundary mesh location. Renamed wall boundary class to WallSitesMeshMapping. Now outputs boundary mesh location.
