+++ CMakeLists.txt
@@ -10,6 +10,7 @@ SET(CXX_SRC
     src/CohesionStressRHS.cpp
     src/PlateletSource.cpp
     src/utility_functions.cpp
+    src/WallSitesMeshMapping.cpp
     )
 
 ADD_LIBRARY(Clot)
