+++ include/clot/BoundaryMeshMapping.h
@@ -32,7 +32,7 @@ public:
     /*!
      * \brief Default deconstructor.
      */
-    ~BoundaryMeshMapping();
+    virtual ~BoundaryMeshMapping() = default;
 
     /*!
      * \brief Default constructor.
@@ -49,45 +49,15 @@ public:
      */
     BoundaryMeshMapping& operator=(const BoundaryMeshMapping& that) = delete;
 
-    /*!
-     * \brief Set the initial number of wall sites.
-     */
-    void setInitialConditions();
-
-    void initializeFEData() override;
-
-    void initializeEquationSystems() override;
-
     void buildBoundaryMesh() override;
 
     void updateBoundaryLocation(double time, unsigned int part, bool end_of_timestep = false) override;
 
-    /*!
-     * \brief Spread the number of used wall sites. The provided patch index must have enough ghost cell width to
-     * perform the requested spreading operation.
-     */
-    void spreadWallSites(int w_idx = IBTK::invalid_index);
-
-    /*!
-     * \brief Return the patch index containing the wall sites. The Eulerian description of the wall sites are updated
-     * automatically when the boundary mesh moves.
-     */
-    inline int getWallSitesPatchIndex()
-    {
-        return d_w_idx;
-    }
-
 protected:
-    std::string d_W_sys_name = "W_SYSTEM";
     IBTK::FEDataManager* d_base_data_manager;
 
     // We need a map between original mesh nodes and bdry mesh nodes
     std::map<libMesh::Node*, libMesh::Node*> d_bdry_base_node_map;
-
-    // We also maintain an Eulerian description of the wall sites
-    SAMRAI::tbox::Pointer<SAMRAI::pdat::CellVariable<NDIM, double>> d_w_var;
-    int d_w_idx = IBTK::invalid_index;
-
 private:
     void commonConstructor(SAMRAI::tbox::Pointer<SAMRAI::tbox::Database> input_db);
 };
